TESTS: D3-mac Add tools for validating behavior
Created a test structure where emitted logs of a movie can be recorded
and validated on playback in ScummVM.

It's now possible to create a movie that calls scummVMLog with
interesting log lines by using a prepared Shared Cast (D3-mac-record).
Over on ScummVM the same movie can be played with a different Shared
Cast (D3-mac-replay). This replay shared cast will check if
the log emitted is identical to the log created on run.